Reply change int to long

+++ features/user/chat_common/src/main/java/com/tokopedia/chat_common/domain/mapper/GetExistingChatMapper.kt
@@ -190,7 +190,6 @@ open class GetExistingChatMapper @Inject constructor() {
 
         val variant: List<AttachmentVariant> = pojoAttribute.productProfile.variant ?: emptyList()
 
-        //TODO: remove toLong()
         if (pojoAttribute.isBannedProduct()) {
             return BannedProductAttachmentViewModel(
                     chatItemPojoByDateByTime.msgId.toString(),
@@ -210,7 +209,7 @@ open class GetExistingChatMapper @Inject constructor() {
                     chatItemPojoByDateByTime.msg,
                     canShowFooterProductAttachment(chatItemPojoByDateByTime.isOpposite,
                             chatItemPojoByDateByTime.role),
-                    chatItemPojoByDateByTime.blastId.toLong(),
+                    chatItemPojoByDateByTime.blastId,
                     pojoAttribute.productProfile.priceInt,
                     pojoAttribute.productProfile.category,
                     variant,
@@ -249,7 +248,7 @@ open class GetExistingChatMapper @Inject constructor() {
                 chatItemPojoByDateByTime.msg,
                 canShowFooterProductAttachment(chatItemPojoByDateByTime.isOpposite,
                         chatItemPojoByDateByTime.role),
-                chatItemPojoByDateByTime.blastId.toLong(),
+                chatItemPojoByDateByTime.blastId,
                 pojoAttribute.productProfile.priceInt,
                 pojoAttribute.productProfile.category,
                 variant,
